refactor path page
